refactor(site/src/pages/AgentsPage): clean up handleSendFromInput and co-locate PendingAttachment
Refactor the unwieldy optimistic-clear/recovery block in
handleSendFromInput into two named helpers:

- clearInputForHistoryEdit(message): clears the composer for an
  in-flight history edit and returns a rollback closure that
  restores the editing draft on failure.

- finalizeSuccessfulSend(editedMessageID, queueEditID): clears
  all input and editing state after a successful send.

The main handleSendFromInput function is now ~20 lines of clear
flow instead of ~55 lines of mixed concerns.

Move the PendingAttachment type from the standalone types.ts file
into ChatPageContent.tsx where it is produced (in the onSend
callback's attachment collection logic) and delete types.ts.
