--- site/src/pages/AgentsPage/types.ts
@@ -1,4 +1,0 @@
-export type PendingAttachment = {
-	fileId: string;
-	mediaType: string;
-};
